Bundle filterlist in JS
